2024/02/19 경록

+++ src/main/java/com/preschool/preschoolhome/entity/MemoryAlbumEntity.java
@@ -10,7 +10,7 @@ public class MemoryAlbumEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "imemory_pic", nullable = false)
-    private Long imemoryPic;
+    private Integer imemoryPic;
 
     @ManyToOne
     @JoinColumn(name = "imemory", nullable = false)
